Hack on particles.js
Wrap most of the code into a function that only gets called on the
client side. Export the return value of this function on the
client side, and export a noop on the server.

Move the polyfills into an if statement that checks for the
existance of window.

Change some function declarations inside if statements/switches to
expressions.

Rename to isoparticles.js and move into lib.